KH2: Deathlink and ingame item popups (#5206)
---------
