+++ worlds/kh2/ClientStuff/WorldLocations.py
@@ -1,5 +1,5 @@
 import typing
-from .Names import LocationName
+from ..Names import LocationName
 
 
 class WorldLocationData(typing.NamedTuple):
@@ -895,6 +895,87 @@ tornPageLocks = {
     "TornPage3": WorldLocationData(0x1DB8, 2),  # --Scenario_3_start
     "TornPage4": WorldLocationData(0x1DB8, 4),  # --Scenario_4_start
     "TornPage5": WorldLocationData(0x1DB8, 7),  # --Scenario_5_start
+}
+# (Room,Event) : Boss event
+# Used for deathlink to say Sora died at : (Room,Event) which returns like Old pete
+DeathLinkPair = {
+    (2, 34, 157): "Twilight Thorn",
+    (2, 5, 88):   "Setzer",
+    (2, 20, 137): "Axel",
+    (2, 4, 80):   "Sandlot",
+    (2, 41, 186): "Mansion fight",
+    (2, 40, 161): "Betwixt and Between",
+    (2, 20, 213): "Data Axel",
+    (4, 8, 52):   "Bailey",
+    (4, 20, 86):  "Corridor",
+    (4, 18, 73):  "Dancers",
+    (4, 4, 55):   "HB Demyx",
+    (4, 16, 65):  "FF Cloud",
+    (4, 17, 66):  "1k Heartless",
+    (4, 1, 75):   "Sephiroth",
+    (4, 4, 114):  "Data Demyx",
+    (5, 11, 72):  "Thresholder",
+    (5, 3, 69):   "Beast",
+    (5, 5, 79):   "Dark Thorn",
+    (5, 4, 74):   "Dragoons",
+    (5, 15, 82):  "Xaldin",
+    (5, 15, 97):  "Data Xaldin",
+    (6, 7, 114):  "Cerberus",
+    (6, 17, 123): "OC Demyx",
+    (6, 8, 116):  "OC Pete",
+    (6, 18, 171): "Hydra",
+    (6, 6, 126):  "Auron Statue fight",
+    (6, 19, 202): "Hades",
+    (4, 34, 151): "Zexion", # all as fights are in hb "world"
+    (7, 9, 2):    "Abu",
+    (7, 13, 79):  "Chasm fight",
+    (7, 10, 58):  "Treasure Room",
+    (7, 3, 59):   "Lords",
+    (7, 14, 100): "Carpet",
+    (7, 5, 62):   "Genie Jafar",
+    (4, 33, 142): "Lexaeus",
+    (8, 5, 72):   "Cave",
+    (8, 7, 73):   "Summit",
+    (8, 9, 75):   "Shan Yu",
+    (8, 10, 78):  "Antechamber fight",
+    (8, 8, 79):   "Storm Rider",
+    (10,14, 55):  "Scar",
+    (10,15, 59):  "Groundshaker",
+    (11,2, 63):   "Tutorial",
+    (11,9, 65):   "Ursula's Revenge",
+    (11,4, 55):   "A New Day is Dawning",
+    (12,1, 53):   "Library (DC)",
+    (12,0, 51):   "Minnie Escort",
+    (13,1, 58):   "Old Pete",
+    (13,2, 52):   "Boat Pete",
+    (13,3, 53):   "DC Pete",
+    (4, 38, 145): "Marluxia",
+    (12,7, 67):   "Lingering Will",
+    (14,6, 53):   "Candy Cane Lane fight",
+    (14,3, 52):   "Prison Keeper",
+    (14,9, 55):   "Oogie Boogie",
+    (14,10, 63):  "Presents minigame",
+    (14,7, 64):   "Experiment",
+    (4, 32, 115): "Vexen",
+    (16,2, 55):   "Town",
+    (16,10, 60):  "Barbossa",
+    (16,14, 62):  "Gambler",
+    (16,1, 54):   "Grim Reaper",
+    (17,3, 54):   "Screens",
+    (17,4, 55):   "Hostile Program",
+    (17,7, 57):   "Solar Sailer",
+    (17,9, 59):   "MCP",
+    (4, 33, 143): "Larxene",
+    (18,21, 65):  "Roxas",
+    (18,10, 57):  "Xigbar",
+    (18,14, 58):  "Luxord",
+    (18,15, 56):  "Saix",
+    (18,19, 59):  "Xemnas 1",
+    (18,20, 98):  "Data Xemnas",
+    (18,21, 99):  "Data Roxas",
+    (18,10, 100): "Data Xigbar",
+    (18,15, 102): "Data Saix",
+    (18,14, 101): "Data Luxord",
 }
 all_world_locations = {
     **TWTNW_Checks,
