+++ worlds/kh2/__init__.py
@@ -3,7 +3,8 @@ from typing import List
 
 from BaseClasses import Tutorial, ItemClassification
 from Fill import fast_fill
-from worlds.LauncherComponents import Component, components, Type, launch as launch_component
+from Utils import local_path
+from worlds.LauncherComponents import Component, components, icon_paths, Type, launch as launch_component
 from worlds.AutoWorld import World, WebWorld
 from .Items import *
 from .Locations import *
@@ -16,11 +17,12 @@ from .Subclasses import KH2Item
 
 
 def launch_client():
-    from .Client import launch
+    from .ClientStuff.Client import launch
     launch_component(launch, name="KH2Client")
 
 
-components.append(Component("KH2 Client", func=launch_client, component_type=Type.CLIENT))
+icon_paths['kh2apicon'] = f"ap:{__name__}/data/khapicon.png"
+components.append(Component("KH2 Client", func=launch_client, component_type=Type.CLIENT, icon='kh2apicon'))
 
 
 class KingdomHearts2Web(WebWorld):
@@ -102,16 +104,16 @@ class KH2World(World):
                 self.goofy_ability_dict[ability] -= 1
 
         slot_data = self.options.as_dict(
-            "Goal", 
-            "FinalXemnas", 
-            "LuckyEmblemsRequired", 
-            "BountyRequired",
-            "FightLogic",
-            "FinalFormLogic",
-            "AutoFormLogic",
-            "LevelDepth",
-            "DonaldGoofyStatsanity",
-            "CorSkipToggle"
+                "Goal",
+                "FinalXemnas",
+                "LuckyEmblemsRequired",
+                "BountyRequired",
+                "FightLogic",
+                "FinalFormLogic",
+                "AutoFormLogic",
+                "LevelDepth",
+                "DonaldGoofyStatsanity",
+                "CorSkipToggle"
         )
         slot_data.update({
             "hitlist":                [],  # remove this after next update
@@ -201,6 +203,7 @@ class KH2World(World):
         """
         Determines the quantity of items and maps plando locations to items.
         """
+
         # Item: Quantity Map
         # Example. Quick Run: 4
         self.total_locations = len(all_locations.keys())
