+++ worlds/kh2/docs/en_Kingdom Hearts 2.md
@@ -2,7 +2,7 @@
 
 ## Changes from the vanilla game
 
-This randomizer creates a more dynamic play experience by randomizing the locations of most items in Kingdom Hearts 2. Currently all items within Chests, Popups, Get Bonuses, Form Levels, and Sora's Levels are randomized. This allows abilities that Sora would normally have to be placed on Keyblades with random stats. Additionally, there are several options for ways to finish the game, allowing for different goals beyond beating the final boss.
+This randomizer creates a more dynamic play experience by randomizing the locations of most items in Kingdom Hearts 2. Currently all items within Chests, Popups, Get Bonuses, Form Levels, Summon Levels,and Sora's Levels are randomized. This allows abilities that Sora would normally have to be placed on Keyblades with random stats. Additionally, there are several options for ways to finish the game, allowing for different goals beyond beating the final boss.
 
 ## Where is the options page
 
@@ -39,7 +39,7 @@ In Kingdom Hearts 2, items which need to be sent to other worlds appear in any l
 ## When the player receives an item, what happens?
 
 
-It is added to your inventory. If you obtain magic, you will need to pause your game to have it show up in your inventory, then enter a new room for it to become properly usable.
+It is added to your inventory.
 
 ## What Happens if I die before Room Saving?
 
@@ -80,7 +80,7 @@ The list of possible locations that can contain a bounty:
 - Lingering Will
 - Starry Hill
 - Transport to Remembrance
-- Godess of Fate cup and Hades Paradox cup
+- Goddess of Fate cup and Hades Paradox cup
 
 ## Quality of life:
 
@@ -97,3 +97,27 @@ With the help of Shananas, Num, and ZakTheRobot we have many QoL features such a
 - Removal of Absent Silhouette and go straight into the Data Fights.
 - And much more can be found at [Kingdom Hearts 2 GoA Overview](https://tommadness.github.io/KH2Randomizer/overview/)
 
+## What does each mod do?
+
+1. Archipelago Companion:
+- This mod under the hood is a collection of smaller mods that change things such as item icons and specific things in the GOA that make it work better for archipelago.
+This mod needs to be above the GOA because it has to overwrite the GOA lua script with its own for things to work according to the client. Such as giving you consumable items
+I.E potions, ethers, boosts etc.
+- This mod also has some consistent mods that should always be on such as Port Royal Map Skip, Better STT and Allowing you to enter drive forms where you shouldnt normally be able to (Dive to the Heart before Data Fights, 100 Acre Woods) Credit to KSX on nexus mods for the basis of the script that does this that was changed for it to work in AP.
+- Changes The Absent Silhouettes to be the data version instead of how it works traditionally where you would defeat the absent silhouette version to unlock the data fight using the same entry point.
+- The biggest misconception is that the APCompanion is the client/connects to the server like other game's companions mod. This is not the case. The apcompanion is mainly a collection of static modifications that were taken out of the apworld to reduce seed size.
+- There are many little things this mod does so if you have any questions feel free to ping me (@JaredWeakStrike) in the archipelago discord and I can hopefully answer your question.
+2. TopazTK/ArchipelagoEnablers
+- This is in sense another companion mod and is required for many things to work correctly.
+- Notification System: Allows the client to flip a byte in game for it to trigger a puzzle piece popup/information popup/chest popup
+- Deathlink: Allows the client to flip a byte to kill sora when the client sets sora's hp to 0 (normally it doesnt kill sora when his hp is set to 0)
+- Instant Movement: No need to pause when you obtain movement for it to update.
+- Instant Magic: No need to room transition to update magic
+- Autosave: Creates a save file in slot 99 that is treated like a normal save file. It is made on room transition. Do note: it does overwrite any save in slot 99
+- Soft Reset: All shoulder buttons+start. For ds4 its L1+l2+R1+R2+Options for example
+3. TopazTK/ArchipelagoEnablersLITE
+- Everything in Archipelago Enablers EXCEPT auto save and soft reset
+- This mod is to be used with H2FM-Mods-equations19/auto-save or H2FM-Mods-equations19/soft-reset
+- Both equations' mods require KH2FM-Mods-equations19/KH2-Lua-Library
+- equations19/soft-reset: Use all shoulder buttons+start to reset 
+- equations19/auto-save: To load an auto-save, hold down the Select or your equivalent on your preferred controller while choosing a NO PROGRESS save file i.e. A save file that is at the start of the game and has no progress made. Make sure to hold the button down the whole time.
